Update OnSelectedItemChanged to use Text property if it's not empty.
If the Text and ValueMemberPath properties are both bound in XAML without a SelectedItem being set (e.g. the Text property contains free text), when OnSelectedItemChanged is triggered by ValueMemberPath binding event it will result in an empty Text property being set because SearchText is an empty string on load.

This change updates it so that OnSelectedItemChanged will take into account the Text property containing a value and use that before using SearchText.